fix: fsutils AvailableSpace

+++ internal/fsutils/fsutils_unix.go
@@ -14,5 +14,5 @@ func AvailableSpace(path string) (uint64, error) {
 	if err != nil {
 		return 0, err
 	}
-	return stat.Bavail * uint64(stat.Bsize), nil
+	return uint64(stat.Bavail) * uint64(stat.Bsize), nil
 }
